requirements.txt
streamlit
langchain
langchain_google_genai
langchain_community
faiss-cpu
python-dotenv
